+++ tests/fixtures/noqa.py
@@ -327,7 +327,10 @@ print(one == 'a' or one == 'b')  # noqa: WPS514
 file_obj = open('filaname.py')  # noqa: WPS515
 print(type(file_obj) == int)  # noqa: WPS516
 
-print(*[])  # noqa: WPS517
+print(*[], **{'@': 1})  # noqa: WPS517, WPS445
+print(lambda: 0)  # noqa: WPS446
+
+pi = 3.14 # noqa: WPS447
 
 for range_len in range(len(file_obj)):  # noqa: WPS518
     print(range_len)
@@ -559,3 +562,23 @@ print(literal)  # noqa: WPS441
 unhashable = {[]}  # noqa: WPS443
 assert []  # noqa: WPS444
 unhashable = [] * 2  # noqa: WPS435
+
+def literal_none_func(arg: Literal[None]):  # noqa: WPS701
+    """Literal[None]"""
+
+
+def literal_none_return_func() -> Literal[None]:  # noqa: WPS701
+    """Literal[None]"""
+
+
+def nested_annotation_func(arg: Literal[Literal[1, 2], 3]): # noqa: WPS702
+    """Literal[Literal[1, 2], 3]"""
+
+
+def nested_annotation_return_func() -> Union[str, Union[int, float]]: # noqa: WPS702
+    """Union[str, Union[int, float]]"""
+
+
+from json import loads  # noqa: WPS347
+from some_module import a  # noqa: WPS347
+from text import from_file  # noqa: WPS347