+++ tests/test_visitors/test_tokenize/test_comments/test_noqa_count.py
@@ -2,6 +2,7 @@
 
 import pytest
 
+from wemake_python_styleguide.options import defaults
 from wemake_python_styleguide.violations.best_practices import (
     OveruseOfNoqaCommentViolation,
 )
@@ -28,6 +29,38 @@ def test_noqa_overuse(
     file_tokens = parse_tokens('{0}\n'.format(code) * (10 + 1))
 
     visitor = WrongCommentVisitor(default_options, file_tokens=file_tokens)
+    visitor.run()
+
+    assert_errors(visitor, [OveruseOfNoqaCommentViolation])
+
+
+def test_noqa_overuse_is_configurable(
+    parse_tokens,
+    assert_errors,
+    options,
+):
+    """Ensures that `noqa` overuse can be configured by options."""
+    file_tokens = parse_tokens(
+        'wallet = 10  # noqa: WPS002, WPS114\n' * defaults.MAX_NOQA_COMMENTS,
+    )
+
+    options = options(max_noqa_comments=defaults.MAX_NOQA_COMMENTS - 1)
+    visitor = WrongCommentVisitor(options, file_tokens=file_tokens)
+    visitor.run()
+
+    assert_errors(visitor, [OveruseOfNoqaCommentViolation])
+
+
+def test_noqa_comments_can_be_forbidden(
+    parse_tokens,
+    assert_errors,
+    options,
+):
+    """Ensures that `noqa` comments can be turned off completely."""
+    file_tokens = parse_tokens('wallet = 10  # noqa: WPS002, WPS114')
+
+    options = options(max_noqa_comments=0)
+    visitor = WrongCommentVisitor(options, file_tokens=file_tokens)
     visitor.run()
 
     assert_errors(visitor, [OveruseOfNoqaCommentViolation])