+++ wemake_python_styleguide/presets/types/tree.py
@@ -28,6 +28,7 @@ PRESET: Final = (
     statements.StatementsWithBodiesVisitor,
     statements.WrongParametersIndentationVisitor,
     statements.PointlessStarredVisitor,
+    statements.WrongNamedKeywordVisitor,
 
     keywords.WrongRaiseVisitor,
     keywords.WrongKeywordVisitor,
@@ -42,6 +43,7 @@ PRESET: Final = (
 
     attributes.WrongAttributeVisitor,
     annotations.WrongAnnotationVisitor,
+    annotations.WrongNestedAnnotationVisitor,
 
     functions.WrongFunctionCallVisitor,
     functions.FunctionDefinitionVisitor,
@@ -58,7 +60,7 @@ PRESET: Final = (
     naming.WrongVariableAssignmentVisitor,
     naming.WrongVariableUsageVisitor,
 
-    builtins.MagicNumberVisitor,
+    builtins.WrongNumberVisitor,
     builtins.WrongStringVisitor,
     builtins.WrongAssignmentVisitor,
     builtins.WrongCollectionVisitor,